compleated functionality of profile picture upload
